requirements.txt
flask==3.0.3
anthropic==0.29.0
python-dotenv==1.0.1
